refactor: merge the nodeplotlib files into one library
