+++ apps/dev-server/src/data/2d-histogram.ts
@@ -1,4 +1,4 @@
-import { Layout, Plot } from '@npl/interfaces';
+import { Layout, Plot } from '@npl/nodeplotlib';
 
 function normal() {
   let x = 0,
