+++ apps/dev-server/src/data/scatter.ts
@@ -1,4 +1,4 @@
-import { Plot } from '@npl/interfaces';
+import { Plot } from '@npl/nodeplotlib';
 
 const trace1: Plot = {
   x: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
