--- apps/nodeplotlib/src/environments/environment.ts
@@ -1,3 +1,0 @@
-export const environment = {
-  production: false,
-};
