--- apps/nodeplotlib/src/main.ts
@@ -1,1 +1,0 @@
-export { plot, Plot, Layout } from '@npl/core';
