--- apps/nodeplotlib/src/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "nodeplotlib",
-  "version": "0.8.0-rc0",
-  "license": "MIT",
-  "main": "main.js",
-  "types": "main.d.ts",
-  "files": [
-    "main.js",
-    "main.d.ts",
-    "web/**"
-  ],
-  "dependencies": {
-    "@types/plotly.js": "^1.54.17"
-  },
-  "peerDependencies": {
-    "rxjs": "^6.5.3"
-  }
-}
