--- apps/nodeplotlib/webpack.config.js
@@ -1,15 +1,0 @@
-const CopyPlugin = require('copy-webpack-plugin');
-
-module.exports = (x) => ({
-  ...x,
-  plugins: [
-    ...x.plugins,
-    new CopyPlugin({
-      patterns: [
-        { from: 'dist/apps/web', to: 'web' },
-        { from: 'apps/nodeplotlib/src/main.d.ts', to: '' },
-        { from: 'LICENSE', to: '' },
-      ],
-    }),
-  ],
-});
