--- libs/core/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@npl/core",
-  "version": "0.8.0-rc0"
-}
