--- libs/core/src/index.ts
@@ -1,2 +1,0 @@
-export { bootstrap, plot } from './lib/nodeplotlib';
-export { Plot, Layout } from '@npl/interfaces';
