--- libs/interfaces/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": [["@nrwl/web/babel", { "useBuiltIns": "usage" }]]
-}
