--- libs/interfaces/README.md
@@ -1,7 +1,0 @@
-# interfaces
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test interfaces` to execute the unit tests via [Jest](https://jestjs.io).
