--- libs/interfaces/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@npl/interfaces",
-  "version": "0.0.1"
-}
