--- libs/interfaces/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib';
