+++ libs/nodeplotlib/.babelrc
