+++ libs/nodeplotlib/README.md
@@ -1,0 +1,7 @@
+# nodeplotlib
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test nodeplotlib` to execute the unit tests via [Jest](https://jestjs.io).
