+++ libs/nodeplotlib/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'core',
+  displayName: 'nodeplotlib',
   preset: '../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -11,5 +11,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/libs/core',
+  coverageDirectory: '../../coverage/libs/nodeplotlib',
 };
