+++ libs/nodeplotlib/src/index.ts
@@ -1,0 +1,6 @@
+// Exports required for the actual nodeplotlib
+export { Plot, Layout } from './lib/interfaces/plot';
+export { plot } from './lib/nodeplotlib';
+
+// Exports required for the web app
+export { PlotData } from './lib/interfaces/plot';
