+++ libs/nodeplotlib/src/lib/interfaces/index.ts
