+++ libs/nodeplotlib/src/lib/open-window.ts
