+++ libs/nodeplotlib/src/lib/server/services/bridge.service.ts
