+++ libs/nodeplotlib/tsconfig.spec.json
@@ -6,14 +6,14 @@
     "types": ["jest", "node"]
   },
   "include": [
-    "**/*.spec.ts",
     "**/*.test.ts",
-    "**/*.spec.tsx",
+    "**/*.spec.ts",
     "**/*.test.tsx",
-    "**/*.spec.js",
+    "**/*.spec.tsx",
     "**/*.test.js",
-    "**/*.spec.jsx",
+    "**/*.spec.js",
     "**/*.test.jsx",
+    "**/*.spec.jsx",
     "**/*.d.ts"
   ]
 }
